% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeFilterAndCollapse_v5.R
\name{MergeFilterAndCollapse}
\alias{MergeFilterAndCollapse}
\title{'MergeFilterAndCollapse'}
\usage{
MergeFilterAndCollapse(
  listdatasetL,
  datasetS,
  key,
  condition,
  saveintermediatedataset = F,
  nameintermediatedataset,
  additionalvar,
  sorting,
  strata,
  summarystat,
  typemerge = 1
)
}
\arguments{
\item{listdatasetL}{a list of one or more data.table() datasets, containing multiple records per -key-. In case the list contains more than one dataset, make sure the names of the -key- variables are equal across datasets.}

\item{datasetS}{(optional) a data.table() dataset, containing one record per -key-}

\item{key}{a vector containing the name(s) of the column identifying the units of observations in -listdatasetL- and in -datasetS-. If the units of observations are identified by variables with the same name, the name can be listed just once, otherwise list first the name in -listdatasetL- and second the name in -datasetS-.}

\item{condition}{(optional) a string containing a condition on the rows of the product between -datasetS- and-datasetL-. Only rows of the product that comply with the condition will be further processed by the function.}

\item{saveintermediatedataset}{(optional) a logical parameter, by default set to FALSE. If it is TRUE the intermediate dataset obtained after -listdatasetL- is merged with -datasetS- and filtered with -condition- will be saved. If -additionalvar-is specified, the intermediate dataset will also contain the new variables. If -nameintermediatedataset- is not specified, the intermedate dataset is saved in the working directory with name 'intermediatedataset'.}

\item{nameintermediatedataset}{(optional) a string specifying the namefile of the intermediate dataset  (path is comprised in the name,if any).}

\item{additionalvar}{(optional) a  list of lists containing additional variables to be created on the merged dataset before computing summary statistics. Each list is made up of three parts: the first is the name to give to the new variable, the second is the content of the variable, and the third is an optional condition filtering the rows to fill.}

\item{sorting}{(optional) a vector containing the variable(s) the dataset must be sorted by before computing summary statistics}

\item{strata}{a vector of column name(s) of -datasetS- or/and -datasetL-  and/or -additionalvar- across which the dataset is collapsed.}

\item{summarystat}{a list of lists each one containing three elements: first a summary statistic to be computed (values allowed are:  mean, min, max, sd, mode, first, second, secondlast, last, exist, sum, count), on which variable to computed it and optionally, as third element with the new name to give to the new variable.}

\item{typemerge}{(optional) a dichotomous parameter: 1 (default) indicates a one-to-many merge, 2 a many-to-many merge.}
}
\description{
The function MergeFilterAndCollapse performs the merge between a dataset with one row per unit of observation and a dataset with multiple rows while filtering by a condition. Furthermore it is possibile to collapse and compute summary statistics across strata of a categorical variable.
}
\details{
The  function MergeFilterAndCollapse operates several steps:
1) Merge -listdatasetL- with -datasetS- per -key- while filtering with -condition-. The merge may be one-to-many (default) or many-to-many (if -typemerge- is set to 2). As an option, -datasetS- may be missing, in this case only -listdatasetL- is filtered.
2) If the parameter -additionalvar- is specified, additional variables can be computed on the merged dataset and then used in the next step to compute summart statistics. This intermediate dataset may be saved for later processing
3) The merged dataset is collapsed across strata of -strata- and rows summary statistics specified in the parameter -summarystat- are computed. The possible value are: minimum (write "min"), maximum ("max"), mean ("mean"), standard deviation ("sd"), mode ("mode"), first element("first"), second element ("second"), second last element ("secondlast"),an element exist ("exist"),sum ("sum"), count ("count").
MergeFilterAndCollapse returns as an output a data.table() dataset with one row for each level of the strata variable(s) and as columns: the strata variable(s), a column for each specified summary statistic.
}
